+++ src/main/kotlin/com/swiftmako/jormanager/model/NativeAsset.kt
@@ -1,7 +1,9 @@
 package com.swiftmako.jormanager.model
 
+import java.math.BigInteger
+
 data class NativeAsset(
         val name: String,
         val policy: String,
-        val amount: Long,
+        val amount: BigInteger,
 )
